# schema/openapi/openapi.yaml
openapi: '3.0.3'
info:
  title: Backbone API
  version: '1.0.0'
  description: |
    Auto-generated by backbone-cli.
    This API provides 11 standard CRUD endpoints for each model.
servers:
  - url: http://localhost:3000
    description: Development server
...
paths:
  /api/v1/accounts:
    get:
      tags:
        - Accounts
      summary: List Accounts
      description: Get a paginated list of Accounts with optional filtering and sorting
      operationId: listAccounts
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/PerPageParam'
        - $ref: '#/components/parameters/SortByParam'
        - $ref: '#/components/parameters/SortOrderParam'
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AccountList'
        '401':
          $ref: '#/components/responses/Unauthorized'

    post:
      tags:
        - Accounts
      summary: Create Account
      description: Create a new Account
      operationId: createAccount
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateAccountRequest'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Account'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
...
  /api/v1/accounting_posts:
    get:
      tags:
        - AccountingPosts
      summary: List AccountingPosts
      description: Get a paginated list of AccountingPosts with optional filtering and sorting
      operationId: listAccountingPosts
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/PerPageParam'
        - $ref: '#/components/parameters/SortByParam'
        - $ref: '#/components/parameters/SortOrderParam'
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AccountingPostList'
        '401':
          $ref: '#/components/responses/Unauthorized'

    post:
      tags:
        - AccountingPosts
      summary: Create AccountingPost
      description: Create a new AccountingPost
      operationId: createAccountingPost
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateAccountingPostRequest'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AccountingPost'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
...
components:
  schemas:
    Account:
      type: object
      required:
        - id
        - company_id
        - account_number
        - account_code
        - name
        - account_type
        - account_subtype
        - normal_balance
        - level
        - is_header
        - is_detail
        - currency
        - opening_balance
        - current_balance
        - is_reconcilable
        - has_budget
        - sort_order
        - show_in_reports
        - status
        - is_system
        - is_cloned
        - metadata
      properties:
        id:
          type: string
          format: uuid
        company_id:
          type: string
          format: uuid
        account_number:
          type: string
          maxLength: 20
        account_code:
          type: string
          maxLength: 30
        name:
          type: string
          maxLength: 100
        description:
          nullable: true
          type: string
          maxLength: 500
        account_type:
          type: string
        account_subtype:
          type: string
        normal_balance:
          type: string
        parent_id:
          nullable: true
          type: string
          format: uuid
        level:
          type: integer
          format: int32
        path:
          nullable: true
          type: string
          maxLength: 255
        is_header:
          type: boolean
        is_detail:
          type: boolean
        currency:
          type: string
        opening_balance:
          type: string
        opening_balance_date:
          nullable: true
          type: string
          format: date
        current_balance:
          type: string
        is_reconcilable:
          type: boolean
        has_budget:
          type: boolean
        budget_amount:
          nullable: true
          type: string
        sort_order:
          type: integer
          format: int32
        show_in_reports:
          type: boolean
        status:
          type: string
        is_system:
          type: boolean
        notes:
          nullable: true
          type: string
          maxLength: 1000
        source_id:
          nullable: true
          type: string
          format: uuid
        is_cloned:
          type: boolean
        chart_code:
          nullable: true
          type: string
          maxLength: 50
        chart_version:
          nullable: true
          type: string
          maxLength: 20
        metadata:
          type: object
...
    CreateAccountRequest:
      type: object
      required:
        - company_id
        - account_number
        - account_code
        - name
        - account_type
        - account_subtype
        - normal_balance
        - level
        - is_header
        - is_detail
        - currency
        - opening_balance
        - current_balance
        - is_reconcilable
        - has_budget
        - sort_order
        - show_in_reports
        - status
        - is_system
        - is_cloned
        - metadata
      properties:
        company_id:
          type: string
          format: uuid
        account_number:
          type: string
          maxLength: 20
        account_code:
          type: string
          maxLength: 30
        name:
          type: string
          maxLength: 100
        description:
          nullable: true
          type: string
          maxLength: 500
        account_type:
          type: string
        account_subtype:
          type: string
        normal_balance:
          type: string
        parent_id:
          nullable: true
          type: string
          format: uuid
        level:
          type: integer
          format: int32
        path:
          nullable: true
          type: string
          maxLength: 255
        is_header:
          type: boolean
        is_detail:
          type: boolean
        currency:
          type: string
        opening_balance:
          type: string
        opening_balance_date:
          nullable: true
          type: string
          format: date
        current_balance:
          type: string
        is_reconcilable:
          type: boolean
        has_budget:
          type: boolean
        budget_amount:
          nullable: true
          type: string
        sort_order:
          type: integer
          format: int32
        show_in_reports:
          type: boolean
        status:
          type: string
        is_system:
          type: boolean
        notes:
          nullable: true
          type: string
          maxLength: 1000
        source_id:
          nullable: true
          type: string
          format: uuid
        is_cloned:
          type: boolean
        chart_code:
          nullable: true
          type: string
          maxLength: 50
        chart_version:
          nullable: true
          type: string
          maxLength: 20
        metadata:
          type: object
...
    AccountList:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/Account'
        meta:
          $ref: '#/components/schemas/PaginationMeta'
    AccountingPost:
      type: object
      required:
        - id
        - company_id
        - source_type
        - source_id
        - posting_type
        - posting_status
        - currency
        - total_debit
        - total_credit
        - retry_count
        - max_retries
        - metadata
      properties:
        id:
          type: string
          format: uuid
        company_id:
          type: string
          format: uuid
        branch_id:
          nullable: true
          type: string
          format: uuid
        source_type:
          type: string
        source_id:
          type: string
          format: uuid
        source_reference:
          nullable: true
          type: string
          maxLength: 100
        journal_id:
          nullable: true
          type: string
          format: uuid
        posting_type:
          type: string
        posting_status:
          type: string
        currency:
          type: string
        total_debit:
          type: string
        total_credit:
          type: string
        scheduled_at:
          nullable: true
          type: string
          format: date-time
        posted_at:
          nullable: true
          type: string
          format: date-time
        failed_at:
          nullable: true
          type: string
          format: date-time
        retry_count:
          type: integer
          format: int32
        max_retries:
          type: integer
          format: int32
        next_retry_at:
          nullable: true
          type: string
          format: date-time
        error_code:
          nullable: true
          type: string
          maxLength: 50
        error_message:
          nullable: true
          type: string
          maxLength: 1000
        reverses_post_id:
          nullable: true
          type: string
          format: uuid
        reversed_by_post_id:
          nullable: true
          type: string
          format: uuid
        posted_by:
          nullable: true
          type: string
          format: uuid
        notes:
          nullable: true
          type: string
          maxLength: 500
        metadata:
          type: object
...
    CreateAccountingPostRequest:
      type: object
      required:
        - company_id
        - source_type
        - source_id
        - posting_type
        - posting_status
        - currency
        - total_debit
        - total_credit
        - retry_count
        - max_retries
        - metadata
      properties:
        company_id:
          type: string
          format: uuid
        branch_id:
          nullable: true
          type: string
          format: uuid
        source_type:
          type: string
        source_id:
          type: string
          format: uuid
        source_reference:
          nullable: true
          type: string
          maxLength: 100
        journal_id:
          nullable: true
          type: string
          format: uuid
        posting_type:
          type: string
        posting_status:
          type: string
        currency:
          type: string
        total_debit:
          type: string
        total_credit:
          type: string
        scheduled_at:
          nullable: true
          type: string
          format: date-time
        posted_at:
          nullable: true
          type: string
          format: date-time
        failed_at:
          nullable: true
          type: string
          format: date-time
        retry_count:
          type: integer
          format: int32
        max_retries:
          type: integer
          format: int32
        next_retry_at:
          nullable: true
          type: string
          format: date-time
        error_code:
          nullable: true
          type: string
          maxLength: 50
        error_message:
          nullable: true
          type: string
          maxLength: 1000
        reverses_post_id:
          nullable: true
          type: string
          format: uuid
        reversed_by_post_id:
          nullable: true
          type: string
          format: uuid
        posted_by:
          nullable: true
          type: string
          format: uuid
        notes:
          nullable: true
          type: string
          maxLength: 500
        metadata:
          type: object
...
    AccountingPostList:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/AccountingPost'
        meta:
          $ref: '#/components/schemas/PaginationMeta'
...
    PaginationMeta:
      type: object
      properties:
        total:
          type: integer
          example: 100
        page:
          type: integer
          example: 1
        per_page:
          type: integer
          example: 20
        total_pages:
          type: integer
          example: 5
    Error:
      type: object
      required:
        - code
        - message
      properties:
        code:
          type: string
          example: VALIDATION_ERROR
        message:
          type: string
          example: Invalid input data
        details:
          type: object
          additionalProperties: true
  parameters:
...
    PageParam:
      name: page
      in: query
      schema:
        type: integer
        minimum: 1
        default: 1
    PerPageParam:
      name: per_page
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
    SortByParam:
      name: sort_by
      in: query
      schema:
        type: string
    SortOrderParam:
      name: sort_order
      in: query
      schema:
        type: string
        enum: [asc, desc]
        default: asc
  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
security:
  - bearerAuth: []